// src/shared/api/generated/model/core/notificationViewDto.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * SnapFlow API
 * REST API для SnapFlow
 * OpenAPI spec version: 1.0.0
 */
import type { NotificationViewDtoNotificationType } from './notificationViewDtoNotificationType';

export interface NotificationViewDto {
  id: string;
  userId: string;
  message: string;
  notificationType: NotificationViewDtoNotificationType;
  isRead: boolean;
  createdAt: string;
}